Initial implementation of a work stealing multithreaded solution that will eventually replace the entire games multithreading logic.
The intention is to remove the artifical barriers in the simulation logic where all planets have to complete a step before all planets can start with the next step. Most update logic focuses on a single planet so there is no reason why a specific thread should wait for other planets.

This implementation should completely remove the need for planets to wait for each other which should allow the game to use 100% CPU which it otherwise is not capable of doing.

For now it only multithreads power. There i see a 20-30% performance improvement. I would not expect the same to happen when i add the remaining systems. I would expect a 10-25% performance improvement max solely with this new multithreaded code.

Indirectly through the implementation of this feature i will also have to multithread more logic that previously was not multithreaded. I don't forsee that to be a problem.